Add some unit types and conversion stuff
Also change the way we handle said units. Coordinate system is in Parsecs. Star data (and geometry) is in Solar Radii, however in the Sector View we scale the geometries such that 1 Solar Radius == 1 parsec. Then we set the camera to be 1 AU worth of Solar Radii away, but since the coordinates/units in SceneKit are being treated as Parsecs, it's actually that many parsecs away.